# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Samoey Copilot API
  description: Comprehensive API documentation for Samoey Copilot project
  version: 1.0.0
  contact:
    name: Samoey Copilot Team
    email: support@samoey-copilot.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.samoey-copilot.com/v1
    description: Production server
paths:
...
  /files/upload:
    post:
      tags:
        - Files
      summary: Upload file
      description: Upload a file to the server
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  filename:
                    type: string
                  size:
                    type: integer
                  url:
                    type: string
        '400':
          description: Invalid file or size exceeded
        '401':
          description: Unauthorized
  /files/{file_id}:
    get:
      tags:
        - Files
      summary: Download file
      description: Download a file by ID
      security:
        - BearerAuth: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
        '401':
          description: Unauthorized
  /notifications/:
    get:
      tags:
        - Notifications
      summary: Get notifications
      description: Get user notifications
      security:
        - BearerAuth: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthorized
  /notifications/{notification_id}/read:
    post:
      tags:
        - Notifications
      summary: Mark notification as read
      description: Mark a specific notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
        '404':
          description: Notification not found
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [info, warning, error, success]
        title:
          type: string
        message:
          type: string
        is_read:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
...